flask_api/requirements.txt
Flask==0.10
wsgiref==0.1.2
Flask-MongoKit==0.2
PyMongo==2.8
python-dateutil==2.4